Volunteers can log in and see their computers.
I have added a new field in the host table (username) to know which machines belong to a given user. The data.inc file reflects also the new field in the class host. However, I have to use different names, because the used field for knowing the name of the user is "name", an if I write the same name for the database table field, it will seem to be the name of the host, while it is the name of the user.