Fixed a bug in the voting process.
Before this patch, anyone will be able to change the active projects of Jarifa by simply visiting the web page: http://domain/jarifa/votes.php. With the patch, it is required a md5 has of a password for the projects, so to change the projects you have to provide a hash in the following way:
http://domain/jarifa/votes.php?auth=HASH

Recovered back also the messages.po file.