Added an invisible captcha to the Join Us page.
The page presents a form, where there is a field form that will not be visible on standard browsers, except for bots and users using text browsers. The field should be empty, as it is not shown for users, so if a bot fields it, it will be treated as a spam bot, trashing the request.