Fixed CDATA opaque parse problem.
The opaque data is surrounded by a CDATA tag that was ignored by the RPC causing to not authenticate the host and create a
new entry in the DB each time a RPC was issued by the client. This commit fixes the problem, requiring to have at least a
BOINC 6.10.x version.